+++ test/sparql_client/query/construct_test.exs
@@ -36,7 +36,7 @@ defmodule SPARQL.Client.Query.ConstructTest do
 
     assert_equal_graph(
       @result_graph,
-      SPARQL.Client.query(@example_query, @example_endpoint, result_format: :turtle)
+      SPARQL.Client.construct(@example_query, @example_endpoint, result_format: :turtle)
     )
   end
 
@@ -52,7 +52,7 @@ defmodule SPARQL.Client.Query.ConstructTest do
         }
     end)
 
-    assert SPARQL.Client.query(@example_query, @example_endpoint, result_format: :jsonld) ==
+    assert SPARQL.Client.construct(@example_query, @example_endpoint, result_format: :jsonld) ==
              {:ok, @result_dataset}
   end
 
@@ -74,7 +74,7 @@ defmodule SPARQL.Client.Query.ConstructTest do
 
     assert_equal_graph(
       @result_graph,
-      SPARQL.Client.query(@example_query, @example_endpoint,
+      SPARQL.Client.construct(@example_query, @example_endpoint,
         result_format: :ntriples,
         request_method: :get,
         protocol_version: "1.1"
@@ -96,7 +96,7 @@ defmodule SPARQL.Client.Query.ConstructTest do
         }
     end)
 
-    assert SPARQL.Client.query(@example_query, @example_endpoint,
+    assert SPARQL.Client.construct(@example_query, @example_endpoint,
              result_format: :nquads,
              request_method: :post,
              protocol_version: "1.1"
@@ -121,7 +121,7 @@ defmodule SPARQL.Client.Query.ConstructTest do
 
       assert_equal_graph(
         @result_graph,
-        SPARQL.Client.query(@example_query, @example_endpoint)
+        SPARQL.Client.construct(@example_query, @example_endpoint)
       )
     end
 
@@ -135,7 +135,7 @@ defmodule SPARQL.Client.Query.ConstructTest do
           }
       end)
 
-      assert SPARQL.Client.query(@example_query, @example_endpoint) ==
+      assert SPARQL.Client.construct(@example_query, @example_endpoint) ==
                {:error,
                 "SPARQL service responded with text/plain content which can't be interpreted. Try specifying one of the supported result formats with the :result_format option."}
     end
@@ -154,7 +154,7 @@ defmodule SPARQL.Client.Query.ConstructTest do
 
       assert_equal_graph(
         @result_graph,
-        SPARQL.Client.query(@example_query, @example_endpoint, result_format: :turtle)
+        SPARQL.Client.construct(@example_query, @example_endpoint, result_format: :turtle)
       )
     end
   end
@@ -174,7 +174,7 @@ defmodule SPARQL.Client.Query.ConstructTest do
 
       assert_equal_graph(
         @result_graph,
-        SPARQL.Client.query(@example_query, @example_endpoint,
+        SPARQL.Client.construct(@example_query, @example_endpoint,
           headers: %{"Accept" => "text/turtle"}
         )
       )
@@ -192,7 +192,7 @@ defmodule SPARQL.Client.Query.ConstructTest do
           }
       end)
 
-      assert SPARQL.Client.query(@example_query, @example_endpoint,
+      assert SPARQL.Client.construct(@example_query, @example_endpoint,
                headers: %{"Accept" => "text/plain"}
              ) ==
                {:error,
@@ -213,7 +213,7 @@ defmodule SPARQL.Client.Query.ConstructTest do
 
       assert_equal_graph(
         @result_graph,
-        SPARQL.Client.query(@example_query, @example_endpoint,
+        SPARQL.Client.construct(@example_query, @example_endpoint,
           result_format: :ntriples,
           headers: %{"Accept" => "text/plain"}
         )